% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_crps.R
\name{calc_crps}
\alias{calc_crps}
\title{calculate continuous rank probability score from a VETS model forecast distribution}
\usage{
calc_crps(simulation, y_test)
}
\arguments{
\item{simulation}{\code{matrix}. The forecast object, which can either be a simulation from
\code{\link[tsvets]{predict.tsvets.estimate}} or a list of reconciled distributions from \code{\link{thief_vets}}}

\item{y_test}{\code{xts matrix}. The out-of-sample test horizon of the outcome series}
}
\value{
a \code{summary} of the out-of-sample CRPS
}
\description{
calculate continuous rank probability score from a VETS model forecast distribution
}
\details{
The sum of the CRPS is calculated against the test observations in \code{y_test} and returned. Calculations
use the \code{\link[scoringRules]{crps_sample}} function
}
