% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{copula_params}
\alias{copula_params}
\title{For discrete series, Probability Integral Transform sampling maps observations onto a Uniform distribution that is then mapped
onto a Gaussian distribution for modelling. This is essentially a copula method where we assume the 'marginal'
is a discrete distribution. This function determines whether a Poisson or Negative Binomial is more appropriate for the series
and then generates the approximately Gaussian PIT-transformed values of the series.}
\usage{
copula_params(y, non_neg = TRUE, censor = 1, k = 1)
}
\arguments{
\item{y}{\code{ts} object containing the discrete time series. \code{NA}s are allowed and will be interpolated using a combination of
\code{\link[zoo]{rollmean}} and \code{\link[forecast]{na.interp}}}

\item{non_neg}{\code{logical} indicating whether the series is restricted to be non-negative. Default is \code{TRUE}}

\item{censor}{\code{numeric} value ranging \code{0 - 1} indicating the upper quantile of values to truncate high outliers to
prior to estimating discrete distribution parameters. Useful when large outliers an lead to inflated estimates of
the distribution mean. Default is \code{0.99}}

\item{k}{\code{integer} indicating the  width of the rolling window to use for smoothly interpolating missing values. See more in
\code{\link[zoo]{rollmean}}}
}
\value{
A \code{list} containing the original series with missing values interpolated, the approximately Gaussian transformed
series and the estimated discrete distribution parameter(s)
}
\description{
For discrete series, Probability Integral Transform sampling maps observations onto a Uniform distribution that is then mapped
onto a Gaussian distribution for modelling. This is essentially a copula method where we assume the 'marginal'
is a discrete distribution. This function determines whether a Poisson or Negative Binomial is more appropriate for the series
and then generates the approximately Gaussian PIT-transformed values of the series.
}
