% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fc_jags_mv.R
\name{fc_jags_mv}
\alias{fc_jags_mv}
\title{Fit a Bayesian forecast model to a multivariate set of discrete time series}
\usage{
fc_jags_mv(
  y,
  model = "arima",
  horizon,
  frequency,
  use_nb = TRUE,
  n.chains = 2,
  n.adapt = 1000,
  n.burnin = 10000,
  n.iter = 10000,
  thin = 10,
  auto_update = TRUE
)
}
\arguments{
\item{y}{\code{xts matrix}. The discrete outcome series to be modelled. \code{NAs} are allowed}

\item{model}{\code{character} either 'TBATS', 'arima' or 'stlf'}

\item{horizon}{\code{integer}. The horizon to forecast. Defaults to \code{frequency}}

\item{frequency}{\code{integer}. The seasonal frequency in \code{y}}

\item{use_nb}{\code{logical} If \code{TRUR}, use a Negative Binomial likelihood with estimated
overdispersion parameter r;}

\item{n.chains}{\code{integer} the number of parallel chains for the model}

\item{n.adapt}{\code{integer} the number of iterations for adaptation. See adapt for details.
If \code{n.adapt} = 0 then no adaptation takes place}

\item{n.iter}{\code{integer} the number of iterations of the Markov chain to run}

\item{auto_update}{\code{logical} If \code{TRUE}, the model is run for up to \code{3} additional sets of
\code{n.iter} iterations, or until the lower 15th percentile of effective sample sizes reaches \code{200}}
}
\value{
A \code{list} containing the posterior forecast from the forecast model and the original fitted
\code{\link[rjags]{jags.model}}
}
\description{
This function takes fitted values from a log-transformed forecast models and uses those as
predictors in a Negative Binomial regression to model the discrete outcome time series. Multivariate
covariance dependencies are incorporated via multivariate normal intercepts
}
\details{
The discrete series are first log-transformed and interpolated to fit univariate forecast models.
Fitted values from the forecast models are then used as predictors in a Negative Binomial Bayesian regression
to model the raw discrete series, while inter-series dependencies are captured using a multivariate normal
intercept
}
