% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_stlf_discreet.R
\name{fit_stlf_discreet}
\alias{fit_stlf_discreet}
\title{Fit an stlf model to a logged version of a discrete time series}
\usage{
fit_stlf_discreet(y, frequency, horizon)
}
\arguments{
\item{y}{\code{xts vector}. The discrete outcome series to be modelled. \code{NAs} are allowed}

\item{frequency}{\code{integer}. The seasonal frequency in \code{y}}

\item{horizon}{\code{integer}. The horizon to forecast. Defaults to \code{frequency}}
}
\value{
A \code{vector} containing fitted values from the model of \code{length(length(y)+horizon)}
}
\description{
This function fits an stlf model with AR errors to a discrete time series that has been interpolated
and log(x + 1/6) transformed
}
