% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_tbats_discreet.R
\name{fit_tbats_discreet}
\alias{fit_tbats_discreet}
\title{Fit a TBATS model to a logged version of a discrete time series}
\usage{
fit_tbats_discreet(y, frequency, horizon)
}
\arguments{
\item{y}{\code{xts vector}. The discrete outcome series to be modelled. \code{NAs} are allowed}

\item{frequency}{\code{integer}. The seasonal frequency in \code{y}}

\item{horizon}{\code{integer}. The horizon to forecast. Defaults to \code{frequency}}
}
\value{
A \code{vector} containing fitted values from the TBATS model of \code{length(length(y)+horizon)}
}
\description{
This function fits a TBATS model to a discrete time series that has been interpolated
and log(x + 1/6) transformed
}
