% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mat2triplet.R
\name{mat2triplet}
\alias{mat2triplet}
\title{Auxiliary function needed by reconcile_thiefnonneg}
\usage{
mat2triplet(x, uniqT = FALSE)
}
\arguments{
\item{x}{\code{matrix}}

\item{uniqT}{\code{logical}}
}
\description{
Auxiliary function needed by reconcile_thiefnonneg
}
