% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{nb_params}
\alias{nb_params}
\title{This function is a wrapper for \code{\link[MASS]{fitdistr}} to generate estimates of Negative Binomial parameters}
\usage{
nb_params(x)
}
\arguments{
\item{x}{\code{vector} of values representing the discrete count vector}
}
\value{
A \code{list} of estimates for the mean and size parameters from a \code{\link[MASS]{fitdistr}} object
}
\description{
This function is a wrapper for \code{\link[MASS]{fitdistr}} to generate estimates of Negative Binomial parameters
}
