% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{paranorm}
\alias{paranorm}
\title{For discrete series, Probability Integral Transform sampling maps observations onto a Uniform distribution that is then mapped
onto a Gaussian distribution for modelling. This is essentially a copula method where we assume the 'marginal'
is a discrete distribution. This function ranks \code{log(x + 0.01)} transformed values of \code{x} to generate
the Uniform(0,1) marginal and then uses \code{\link[stats]{qnorm}} to map these onto an approximately Gaussian
distribution}
\usage{
paranorm(x)
}
\arguments{
\item{x}{\code{vector} of values representing the discrete count vector}
}
\value{
A \code{vector} of approximately Gaussian PIT-transformed values
}
\description{
For discrete series, Probability Integral Transform sampling maps observations onto a Uniform distribution that is then mapped
onto a Gaussian distribution for modelling. This is essentially a copula method where we assume the 'marginal'
is a discrete distribution. This function ranks \code{log(x + 0.01)} transformed values of \code{x} to generate
the Uniform(0,1) marginal and then uses \code{\link[stats]{qnorm}} to map these onto an approximately Gaussian
distribution
}
