% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mv_preds.R
\name{plot_mv_preds}
\alias{plot_mv_preds}
\title{Plot simulation predictions from a reconciled thief model}
\usage{
plot_mv_preds(
  simulation,
  main = "",
  ylab = "Y",
  ylim = c(min(simulation, na.rm = T), max(simulation, na.rm = T))
)
}
\arguments{
\item{simulation}{\code{matrix}. The forecast object, which should be a horizon x simulation \code{matrix}
similar to those returned by \code{\link{thief_vets}}}

\item{main}{\code{character}. Optional label for the plot}

\item{ylab}{\code{character}. Optional label for the y axis}

\item{ylim}{Limits for the y-axis. Default is \code{c(min(simulation, na.rm = T), max(simulation, na.rm = T))}}
}
\value{
A base plot of the forecast medeian (red dashed line) and credible intervals as a gradient polygon
}
\description{
Plot simulation predictions from a reconciled thief model
}
\details{
The forecast distribution is used to calculate highest posterior density credible estimates for the following
intervals: \code{c(0.95, 0.9, 0.85, 0.8, 0.75, 0.7)}
}
