% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_vets_preds.R
\name{plot_vets_preds}
\alias{plot_vets_preds}
\title{Plot simulation predictions from a reconciled VETS model}
\usage{
plot_vets_preds(simulation, main = "", ylab = "Y")
}
\arguments{
\item{simulation}{\code{matrix}. The forecast object, which should be a horizon x simulation \code{matrix}
similar to those returned by \code{\link{thief_vets}}}

\item{main}{\code{character}. Optional label for the plot}

\item{ylab}{\code{character}. Optional label for the y axis}
}
\value{
A base plot of the forecast mean (black line) and 95% prediction intervals (red dashed lines)
}
\description{
Plot simulation predictions from a reconciled VETS model
}
\details{
The sum of the CRPS is calculated against the test observations in \code{y_test} and returned. Calculations
use the \code{\link[scoringRules]{crps_sample}} function
}
