% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{poiss_params}
\alias{poiss_params}
\title{This function is a wrapper for \code{\link[MASS]{fitdistr}} to generate estimates of Poisson parameters}
\usage{
poiss_params(x)
}
\arguments{
\item{x}{\code{vector} of values representing the discrete count vector}
}
\value{
A \code{numeric} estimate for the mean parameter from a \code{\link[MASS]{fitdistr}} object
}
\description{
This function is a wrapper for \code{\link[MASS]{fitdistr}} to generate estimates of Poisson parameters
}
