% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{reconcilethief_restrict}
\alias{reconcilethief_restrict}
\title{Reconcile temporal hierarchical forecasts}
\usage{
reconcilethief_restrict(
  forecasts,
  comb = c("struc", "mse", "ols", "bu", "shr", "sam"),
  mse = NULL,
  residuals = NULL,
  returnall = TRUE,
  aggregatelist = NULL,
  max_agg = NULL,
  nonnegative = FALSE
)
}
\arguments{
\item{forecasts}{List of forecasts. Each element must be a time series of forecasts,
or a forecast object.
The number of forecasts should be equal to k times the seasonal period for each series,
where k is the same across all series.}

\item{comb}{Combination method of temporal hierarchies, taking one of the following values:
\describe{
  \item{"struc"}{Structural scaling - weights from temporal hierarchy}
  \item{"mse"}{Variance scaling - weights from in-sample MSE}
  \item{"ols"}{Unscaled OLS combination weights}
  \item{"bu"}{Bottom-up combination -- i.e., all aggregate forecasts are ignored.}
  \item{"shr"}{GLS using a shrinkage (to block diagonal) estimate of residuals}
  \item{"sam"}{GLS using sample covariance matrix of residuals}
}}

\item{mse}{A vector of one-step MSE values corresponding to each of the forecast series.}

\item{residuals}{List of residuals corresponding to each of the forecast models.
Each element must be a time series of residuals. If \code{forecast} contains a list of
forecast objects, then the residuals will be extracted automatically and this argument
is not needed. However, it will be used if not \code{NULL}.}

\item{returnall}{If \code{TRUE}, a list of time series corresponding to the first argument
is returned, but now reconciled. Otherwise, only the most disaggregated series is returned.}

\item{aggregatelist}{(optional) User-selected list of forecast aggregates to consider}

\item{max_agg}{(optional) \code{integer} specifying the maximum number of temporal aggregation levels
to use when reconciling, via the structural scaling method. Useful if higher levels of aggregation
are unlikely to have 'seen' recent changes in series dynamics and will likely then result in poor
forecasts as a result. Default is \code{NULL}, meaning that all levels of aggregation are used}

\item{nonnegative}{\code{logical} If \code{TRUE}, forecaststs are constrained using non-negative
optimisation to ensure negative forecasts are eliminated. Default is \code{FALSE}}
}
\value{
List of reconciled forecasts in the same format as \code{forecast}.
If \code{returnall==FALSE}, only the most disaggregated series is returned.
}
\description{
A wrapper for \code{\link[thief]{reconcilethief}} that takes forecasts of time series at all levels of temporal aggregation
and combines them using the temporal hierarchical approach of Athanasopoulos et al (2016). Non-negative
optimisation constraints are used when forecasts must not be negative
}
\seealso{
\code{\link{thief}}, \code{\link{tsaggregates}}
}
