% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_mvforecast.R
\name{test_mvforecast}
\alias{test_mvforecast}
\title{Test script to ensure mvforecast is working properly}
\usage{
test_mvforecast(plot = TRUE)
}
\arguments{
\item{plot}{\code{logical}. Default is \code{TRUE}}
}
\value{
A plot from \code{\link{plot_vets_preds}} and a CRPS calculation from \code{\link{calc_crps}}
}
\description{
This function fits a vets model and equivalent thief_vets model on the example ixodes_vets_dat dataset
and returns a plot and a CRPS calculation
}
