% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thief_ensemble.R
\name{thief_ensemble}
\alias{thief_ensemble}
\title{Temporal hierarchy reconciliation of ensemble univariate models}
\usage{
thief_ensemble(
  y,
  k = 1,
  lambda = NULL,
  frequency = 52,
  horizon = NULL,
  cores = parallel::detectCores() - 1,
  max_agg = NULL
)
}
\arguments{
\item{y}{\code{xts matrix}. The outcome series to be modelled. \code{NAs} are currently not supported}

\item{k}{\code{integer} specifying the length of the forecast horizon in multiples of \code{frequency}. Default
is \code{1}, meaning that a final forecast of \code{frequency} horizons will be returned}

\item{lambda}{\code{numeric}. The Box Cox power transformation parameter for all series. Must be
between \code{-1} and \code{2} inclusive. If \code{y_series} contains zeros, \code{lambda} will be set to
\code{max(c(0.7, lambda))} to ensure stability of forecasts}

\item{frequency}{\code{integer}. The seasonal frequency in \code{y}}

\item{horizon}{\code{integer}. The horizon to forecast. Defaults to \code{frequency}}

\item{cores}{\code{integer}. The number of cores to use. This is used to initialize the states of each series
using \code{\link[tsets]{ets_modelspec}}}

\item{max_agg}{(optional) \code{integer} specifying the maximum number of temporal aggregation levels
to use when reconciling, via the structural scaling method. Useful if higher levels of aggregation
are unlikely to have 'seen' recent changes in series dynamics and will likely then result in poor
forecasts as a result. Default is \code{NULL}, meaning that all levels of aggregation are used}
}
\value{
A \code{list} containing the reconciled forecast distributions for each series in \code{y}. Each element in
the \code{list} is a \code{horizon x 1000 matrix} of forecast predictions
}
\description{
This function fits ensemble univariate forecast models on all levels of temporal
aggregation for a multivariate xts timeseries object
}
\details{
Series in \code{y} are aggregated at all possible levels up to annual using \code{\link[thief]{tsaggregates}}.
\code{\link{ensemble_base}} is used on all levels of aggregation to find a weighted ensemble of six
univariate forecast models that minimises mean absolute scaled error. Forecasts are then reconciled
using \code{\link[thief]{reconcilethief}} and are optionally constrained using non-negative optimisation if there are no
negative values in \code{y}. Adjustments to the original unaggregated forecast are incorporated and a distribution of \code{1000} sample
paths for each series' forecast are returned
}
\examples{
\donttest{
library(mvforecast)
data("ixodes_vets_dat")

#Fit a thief_ensemble model
mod1 <- thief_ensemble(y = ixodes_vets_dat$y_train,
frequency = 52, lambda = 1, k = 1,
cores = parallel::detectCores() - 1)

#Calculate the out-of-sample CRPS
calc_crps(mod1, y_test = ixodes_vets_dat$y_test)

Plot simulation results for one of the plots in the NEON dataset
plot_mvforecast(simulation = mod1[[4]])
points(as.vector(ixodes_vets_dat$y_test[,4]))}

}
\references{
Athanasopoulos, G., Hyndman, R.,  Kourentzes, N.,  and Petropoulos, F. Forecasting with temporal hierarchies.
(2017) European Journal of Operational Research 262(1) 60–74
}
\seealso{
\code{\link{ensemble_base}}, \code{\link[forecast]{forecast}},
\code{\link[thief]{reconcilethief}}
}
