% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble_base.R
\name{ensemble_base}
\alias{ensemble_base}
\title{Fit a series of univariate models and return a weighted ensemble forecast}
\usage{
ensemble_base(y, frequency, lambda = NULL, k, bottom_series = FALSE)
}
\arguments{
\item{y}{A \code{ts} object containing the series to forecast}

\item{frequency}{The seasonal frequency of \code{y}}

\item{lambda}{\code{numeric}. The Box Cox power transformation parameter for all series. Must be
between \code{-1} and \code{2} inclusive. If \code{y} contains zeros, \code{lambda} will be set to
\code{max(c(0.7, lambda))} to ensure stability of forecasts}

\item{k}{\code{integer} specifying the length of the forecast horizon in multiples of \code{frequency}}

\item{bottom_series}{\code{logical}. If \code{TRUE}, the two \code{auto.arima} models will be ignored
to ensure bottom level series forecasts are not overconfident}
}
\value{
A \code{list} object containing the ensemble forecast and the ensemble residuals
}
\description{
This function fits nine simple univariate forecast models to the series and calculates weights
that minimise the mean absolute scaled error of a weighted ensemble forecast
}
\details{
A total of nine simple univariate models are tested on the series. These include
\code{\link[forecast]{auto.arima}} with exponentially weighted moving average regressors,
\code{\link[forecast]{auto.arima}} with fourier terms \code{K = 4} regressors,
\code{\link[forecast]{tbats}},
\code{\link[forecast]{ets}},
\code{\link[forecast]{thetaf}},
\code{\link[forecast]{stlm}} with ARIMA errors,
\code{\link[forecast]{rwf}} with drift,
\code{\link[forecast]{naive}} and
\code{\link[forecast]{snaive}}.
The mean absolute scaled error (MASE) of the in-sample
series is calculated for each series, and a weighted mean is optimised to
minimise in-sample MASE using the "L-BFGS-B" algorithm in \code{\link[stats]{optim}}.
}
