% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{transform_fc_preds}
\alias{transform_fc_preds}
\title{For discrete series, Probability Integral Transform sampling maps observations onto a Uniform distribution that is then mapped
onto a Gaussian distribution for modelling. This is essentially a copula method where we assume the 'marginal'
is a discrete distribution. This function takes an approximately Gaussian forecast object
and back-transforms it to the discrete distribution that was originally estimated.}
\usage{
transform_fc_preds(forecast, orig_y, dist_mappings, dist_params)
}
\arguments{
\item{forecast}{\code{forecast} object containing the approximately Gaussian predictions to be back-transformed}

\item{orig_y}{\code{vector} containing the original untransformed y values}

\item{dist_mappings}{\code{vector} containing random draws from the original discrete distribution. This set of draws is used for
estimating rankings of \code{x} and so it should be large enough (\code{N >= 5000}) to represent the full discrete distribution}

\item{params}{the original estimated discrete distribution parameter(s). If the original was Negative Binomially distributed,
\code{params} will be a \code{list} of \code{length} \code{2}.}
}
\value{
A \code{forecast} containing the estimated discrete predictions
}
\description{
For discrete series, Probability Integral Transform sampling maps observations onto a Uniform distribution that is then mapped
onto a Gaussian distribution for modelling. This is essentially a copula method where we assume the 'marginal'
is a discrete distribution. This function takes an approximately Gaussian forecast object
and back-transforms it to the discrete distribution that was originally estimated.
}
